+++ crates/cli/tests/reference/echo.bg.js
@@ -923,7 +923,7 @@ export function __wbg_foo_unwrap(arg0) {
     const ret = Foo.__unwrap(arg0);
     return ret;
 }
-export function __wbindgen_cast_0000000000000000(arg0, arg1) {
+export function __wbindgen_generic_0000000000000000(arg0, arg1) {
     // Cast intrinsic for `Ref(String) -> Externref`.
     const ret = getStringFromWasm0(arg0, arg1);
     return ret;